requirements.txt
langchain>=0.2.12
langchain-core>=0.2.33
langchain-community>=0.2.11
langgraph>=0.2.12
langchain-openai>=0.1.21
openai>=1.47.0
tiktoken>=0.7.0
pandas>=2.2.2
rapidfuzz>=3.9.6
python-dateutil>=2.9.0.post0
pdfplumber>=0.11.4
reportlab>=4.2.2
streamlit>=1.37.1
pydantic>=2.8.2
pydantic-settings>=2.4.0
matplotlib>=3.9.0
pyarrow>=17.0.0
